# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Drawer",
    "description": "Drawer Application API"
  },
  "host": "127.0.0.1:8080",
  "basePath": "/api",
...
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/users/register": {
      "post": {
        "tags": ["POST Service"],
        "summary": "Create/Register a new user",
        "parameters": [
          {
            "in": "body",
            "name": "scribble",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "description": "min 1, max 50"
                },
                "email": {
                  "type": "string",
                  "description": "min 6, max 50"
                },
                "password": {
                  "type": "string",
                  "required": true
                },
                "role": {
                  "type": "string",
                  "required": true,
                  "default": "user",
                  "description": "'user' or 'admin'"
                },
                "isLoggedIn": {
                  "type": "boolean",
                  "default": false
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "This service creates a new user"
          }
        }
      }
    },
    "/users/{id}": {
      "delete": {
        "tags": ["Delete Service"],
        "summary": "DELETE a user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "This service deletes an user."
          }
        }
      }
    }
  }
}